+++ docs/source/conf.py
@@ -8,8 +8,18 @@ https://www.sphinx-doc.org/en/master/usage/configuration.html
 | `voxel51.com <https://voxel51.com/>`_
 |
 """
-
+import os
 import re
+import sys
+
+sys.path.insert(0, os.path.abspath("."))
+
+from custom_directives import (
+    CustomButtonDirective,
+    CustomCalloutItemDirective,
+    CustomCardItemDirective,
+)
+from redirects import generate_redirects
 
 import fiftyone.constants as foc
 
@@ -24,12 +34,14 @@ if setup_version != foc.VERSION:
         "and try again." % (setup_version, foc.VERSION)
     )
 
+
 # -- Path setup --------------------------------------------------------------
 
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 #
+
 
 # -- Project information -----------------------------------------------------
 
@@ -49,16 +61,21 @@ extensions = [
     "sphinx.ext.intersphinx",
     "sphinx.ext.napoleon",
     "sphinx.ext.autosectionlabel",
-    "m2r",
     "nbsphinx",
     "sphinx_tabs.tabs",
     "sphinx_copybutton",
+    "autodocsumm",
 ]
 
 # Types of class members to generate documentation for.
-autodoc_default_options = {"members": True, "inherited-members": True}
+autodoc_default_options = {
+    "members": True,
+    "inherited-members": True,
+    "member-order": "bysource",
+    "autosummary": True,
+    "autosummary-no-nesting": True,
+}
 autodoc_inherit_docstrings = True
-autodoc_member_order = "bysource"
 autoclass_content = "class"
 
 # Add any paths that contain templates here, relative to this directory.
@@ -90,6 +107,9 @@ nbsphinx_prolog = """
     :download:`{{ env.doc2path(env.docname, base=None) }} </{{ env.doc2path(env.docname, base=None) }}>`
 
 """
+
+# Path to the redirects file, relative to `source/`
+redirects_file = "redirects"
 
 # -- Options for HTML output -------------------------------------------------
 
@@ -153,3 +173,16 @@ html_context = {
     "link_voxel51_linkedin": "https://www.linkedin.com/company/voxel51/",
     "link_voxel51_twitter": "https://twitter.com/voxel51",
 }
+
+# -- Custom app setup --------------------------------------------------------
+
+
+def setup(app):
+    # Generate page redirects
+    app.add_config_value("redirects_file", "redirects", "env")
+    app.connect("builder-inited", generate_redirects)
+
+    # Custom directives
+    app.add_directive("custombutton", CustomButtonDirective)
+    app.add_directive("customcalloutitem", CustomCalloutItemDirective)
+    app.add_directive("customcarditem", CustomCardItemDirective)